Move controllers into Classroom module
This makes more sense for these controllers, which are only accessed in
the context of a classroom
