+++ app/views/teachers/classrooms/learning_modules/show.html.slim
