+++ config/routes.rb
@@ -45,15 +45,15 @@ Rails.application.routes.draw do
           resources :students, except: [:show] do
             member { post :resend_confirmation }
           end
-        end
 
-        resources :learning_modules, only: :index
-        get "/learning_modules/:id",
-            to: "learning_modules#show",
-            as: :learning_module
-        post "/learning_modules/:id/toggle_present",
-             to: "learning_modules#toggle_present",
-             as: :learning_module_toggle_present
+          resources :learning_modules, only: :index
+          get "/learning_modules/:id",
+              to: "learning_modules#show",
+              as: :learning_module
+          post "/learning_modules/:id/toggle_present",
+               to: "learning_modules#toggle_present",
+               as: :learning_module_toggle_present
+        end
 
         resources :surveys, only: [:index]
         # Unfortunately we can't undo a shallow: true so we need to be verbose here.
