+++ spec/features/teachers/classrooms/supplemental_materials_spec.rb
